Add initial implementation of the TreeScaper-Inference tool.
This submits jobs remotely to the CIPRES REST API to generate bootstrap
trees from sequence inputs.

The input is a collection of PHYLIP files.

Output are the concatenated trees from all finished jobs, the parameters
used to run the job, and the status file managed by CLVTreeScaper2. This
status file can be referenced to verify completion of all jobs, and may
also be provided to CLVTreeScaper2 as an input to resume job submission,
although this version of the wrapper does not support this.